requirements.txt
python-multipart
python-dotenv
mypy
sqlalchemy
uvicorn
pydantic
psycopg2-binary
fastapi